Refactor lots of logic into a PluginStepRegistrationManager DynamicsValue/fake-xrm-easy#122
