+++ src/FakeXrmEasy.Plugins/Pipeline/PipelineProcessor.cs
@@ -1,0 +1,7 @@
+namespace FakeXrmEasy.Pipeline
+{
+    internal class PipelineProcessor
+    {
+        
+    }
+}
