+++ src/FakeXrmEasy.Plugins/Pipeline/PipelineStageExecutionParameters.cs
@@ -1,4 +1,5 @@
-﻿using FakeXrmEasy.Abstractions.Plugins.Enums;
+﻿using System.Collections.Generic;
+using FakeXrmEasy.Abstractions.Plugins.Enums;
 using Microsoft.Xrm.Sdk;
 
 namespace FakeXrmEasy.Pipeline
@@ -20,13 +21,23 @@ namespace FakeXrmEasy.Pipeline
         internal OrganizationResponse Response { get; set; }
 
         /// <summary>
-        /// Snapshot of the entity values before the execution of the request
+        /// Snapshot of the entity values before the execution of a non-bulk operation request
         /// </summary>
         internal Entity PreEntitySnapshot { get; set; }
 
         /// <summary>
-        /// Snapshot of the entity values after the execution of the request
+        /// Snapshot of the entity values after the execution of a non-bulk operation request
         /// </summary>
         internal Entity PostEntitySnapshot { get; set; }
+        
+        /// <summary>
+        /// Snapshot of all the entity preimages before the execution of a bulk operation request
+        /// </summary>
+        internal List<Entity> PreEntitySnapshots { get; set; }
+        
+        /// <summary>
+        /// Snapshot of all the entity postimages after the execution of a bulk operation request
+        /// </summary>
+        internal List<Entity> PostEntitySnapshots { get; set; }
     }
 }
