+++ tests/FakeXrmEasy.Plugins.Tests/Issues/Issue0095.cs
@@ -32,7 +32,7 @@ namespace FakeXrmEasy.Plugins.Tests.Issues
         public void Should_have_all_necessary_entities_for_pipeline_simulation_without_raising_exception()
         {
             _context.Initialize(new DataverseEntities.SdkMessage() { Id = Guid.NewGuid()});
-            var ex = Record.Exception(() => FakeXrmEasy.Pipeline.IXrmFakedContextPipelineExtensions.AddPipelineTypes(_context));
+            var ex = Record.Exception(() => PluginStepRegistrationManager.AddPipelineTypes(_context));
             Assert.Null(ex);
         }
